Refactor: organize code into modules for better structure
- Create lib.rs to manage module organization
- Extract data structures into models.rs module
- Move log loading functions to logs.rs module
- Move Jaeger conversion functions to jaeger.rs module
- Simplify main.rs to only contain main function
- Improve code maintainability and readability
